backend/pyproject.toml
[project]
name = "paper-compass-backend"
version = "0.1.0"
description = "Paper Compass — AI-powered research paper comprehension backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "anthropic>=0.39.0",
    "httpx>=0.28.0",
    "aiolimiter>=1.2.1",
    "lxml>=5.3.0",
    "svgwrite>=1.4.3",
    "python-slugify>=8.0.4",
    "python-multipart>=0.0.12",
    "authlib>=1.3.2",
    "pyjwt[crypto]>=2.10.0",
    "itsdangerous>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
